Fix compile on fedora-34.
Remove a bunch of incompatible throw code.
hopefully doesn't actually change the execution of the code any.

Fix a few compiler warnings, nothing that looked too serious.

+++ contrib/speex/speex-1.0.3/libspeex/stack_alloc.h
@@ -48,7 +48,7 @@
 #else
 
 /*Aligns the stack to a 'size' boundary */
-#define ALIGN(stack, size) ((stack) += ((size) - (int)(stack)) & ((size) - 1))
+#define ALIGN(stack, size) ((stack) += ((size) - (long)(stack)) & ((size) - 1))
 
 /* Allocates 'size' elements of type 'type' on the stack */
 #define PUSH(stack, size, type) (ALIGN((stack),sizeof(type)),(stack)+=((size)*sizeof(type)),(type*)((stack)-((size)*sizeof(type))))
